+++ app/src/main/res/layout/fragment_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.easypick.easypick.fragments.SettingsFragment">
+    tools:context="com.easypick.easypick.fragments.ProfileFragment">
     <LinearLayout
         android:id="@+id/loginInfo"
         android:layout_width="match_parent"
